Rename histogram function.
