+++ src/histogram.js
